#26744 Improvements
