+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/pull/PullServiceImpl.java
@@ -447,6 +447,7 @@ public class PullServiceImpl implements PullService {
             }
         }
 
+        output.info(String.format("%n%n"));
     }
 
 }
