+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/cli/command/language/LanguagePull.java
@@ -5,6 +5,7 @@ import com.dotcms.api.client.pull.language.LanguageFetcher;
 import com.dotcms.api.client.pull.language.LanguagePullHandler;
 import com.dotcms.cli.command.DotCommand;
 import com.dotcms.cli.command.DotPull;
+import com.dotcms.cli.common.CompletePullMixin;
 import com.dotcms.cli.common.OutputOptionMixin;
 import com.dotcms.cli.common.PullMixin;
 import com.dotcms.common.WorkspaceManager;
@@ -49,7 +50,7 @@ public class LanguagePull extends AbstractLanguageCommand implements Callable<In
     static final String LANGUAGE_PULL_MIXIN = "languagePullMixin";
 
     @CommandLine.Mixin
-    PullMixin pullMixin;
+    CompletePullMixin pullMixin;
 
     @CommandLine.Mixin(name = LANGUAGE_PULL_MIXIN)
     LanguagePullMixin languagePullMixin;
